move getAllBlogPageRoutes function to page utils
